Decompose clustering queue processor
